Move `RemoteImageComment` to a separate file
