+++ EssentialFeed/EssentialFeed/Feed API/RemoteImageCommentsLoader.swift
@@ -66,28 +66,6 @@ public class RemoteImageCommentsLoader: ImageCommentsLoader {
 	
 }
 
-public struct RemoteImageComment: Decodable {
-	public let id: UUID
-	public let message: String
-	public let created_at: Date
-	public let author: Author
-	
-	public struct Author: Decodable {
-		public let username: String
-		
-		public init(username: String) {
-			self.username = username
-		}
-	}
-	
-	public init(id: UUID, message: String, created_at: Date, author: RemoteImageComment.Author) {
-		self.id = id
-		self.message = message
-		self.created_at = created_at
-		self.author = author
-	}
-}
-
 private extension Array where Element == RemoteImageComment {
 	func toModels() -> [ImageComment] {
 		map { ImageComment(
